+++ backend/src/main/java/com/tribly/domain/ridetemplate/RideTemplateGroup.java
@@ -33,7 +33,7 @@ public class RideTemplateGroup extends BaseEntity {
 
   @Column(name = "average_speed")
   @Nullable
-  private Integer averageSpeed;
+  private Float averageSpeed;
 
   @Column(name = "max_participants")
   @Nullable
